Next.js SDK doc rewrite
