+++ docs-content/getting-started/fullstack-frameworks/next-js/3_client.md
@@ -1,0 +1,215 @@
+---
+title: Next.js Client
+slug: client
+heading: Next.js Client Instrumentation
+createdAt: 2023-10-03T00:00:00.000Z
+updatedAt: 2023-10-03T00:00:00.000Z
+---
+
+## Installation
+
+```shell
+# with yarn
+npm install @highlight-run/next
+```
+
+## Client Instrumentation
+
+This sections adds session replay and frontend error monitoring to Highlight. This implementation requires React 17 or greater. If you're behind on React versions, follow our [React.js docs](../../3_client-sdk/1_reactjs.md)
+
+1. For the `/pages` directory, you'll want to add `HighlightInit` to `_app.tsx`.
+
+```jsx
+// pages/_app.tsx
+import { AppProps } from 'next/app'
+import CONSTANTS from '../app/constants'
+import { HighlightInit } from '@highlight-run/next/client'
+
+export default function MyApp({ Component, pageProps }: AppProps) {
+	return (
+		<>
+			<HighlightInit
+				// excludedHostnames={['localhost']}
+				projectId={CONSTANTS.NEXT_PUBLIC_HIGHLIGHT_PROJECT_ID}
+				serviceName="my-nextjs-frontend"
+				tracingOrigins
+				networkRecording={{
+					enabled: true,
+					recordHeadersAndBody: true
+				}}
+			/>
+
+			<Component {...pageProps} />
+		</>
+	)
+}
+```
+
+2. For Next.js 13 App Directory, add `HighlightInit` to your `layout.tsx` file.
+
+```jsx
+// src/app/layout.tsx
+import './globals.css'
+
+import CONSTANTS from '../constants'
+import { HighlightInit } from '@highlight-run/next/client'
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+	return (
+		<>
+			<HighlightInit
+				// excludedHostnames={['localhost']}
+				projectId={CONSTANTS.NEXT_PUBLIC_HIGHLIGHT_PROJECT_ID}
+				serviceName="my-nextjs-frontend"
+				tracingOrigins
+				networkRecording={{
+					enabled: true,
+					recordHeadersAndBody: true
+				}}
+			/>
+
+			<html lang="en">
+				<body>{children}</body>
+			</html>
+		</>
+	)
+}
+```
+
+3. Optionally add a React [Error Boundary](https://react.dev/reference/react/Component#catching-rendering-errors-with-an-error-boundary).
+
+```jsx
+// src/app/components/error-boundary.tsx
+'use client'
+
+import { ErrorBoundary as HighlightErrorBoundary } from '@highlight-run/next/client'
+
+export function ErrorBoundary({ children }: { children: React.ReactNode }) {
+	const isLocalhost =
+		typeof window === 'object' && window.location.host === 'localhost'
+
+	return (
+		<HighlightErrorBoundary showDialog={!isLocalhost}>
+			{children}
+		</HighlightErrorBoundary>
+	)
+}
+```
+
+### Skip Localhost tracking
+
+```hint
+We do not recommend enabling this while integrating Highlight for the first time because it will prevent you from validating that your local build can send data to Highlight.
+```
+
+In the case that you don't want local sessions being shipped to Highlight, the `excludedHostnames` prop accepts an array of partial or full hostnames. For example, if you pass in `excludedHostnames={['localhost', 'staging]}`, you'll block `localhost` on all ports, `www.staging.highlight.io` and `staging.highlight.com`.
+
+Alternatively, you could manually call `H.start()` and `H.stop()` to manage invocation on your own.
+
+```jsx
+// src/app/layout.tsx
+<HighlightInit
+	manualStart
+	projectId={CONSTANTS.NEXT_PUBLIC_HIGHLIGHT_PROJECT_ID}
+	serviceName="my-nextjs-frontend"
+/>
+<CustomHighlightStart />
+```
+
+```jsx
+// src/app/custom-highlight-start.tsx
+'use client'
+import { H } from '@highlight-run/next/client'
+
+export function CustomHighlightStart() {
+	useEffect(() => {
+		const shouldStartHighlight = window.location.hostname === 'https://www.highlight.io'
+
+		if (shouldStartHighlight) {
+			H.start();
+
+			return () => {
+				H.stop()
+			}
+		}
+	})
+
+	return null
+}
+```
+
+## Validation
+
+Drop this example component somewhere in your client application to see it in action.
+
+```jsx
+'use client'
+import { useEffect, useState } from 'react'
+import { ErrorBoundary } from 'src/app/components/error-boundary'
+
+export function ErrorButtons() {
+	const [isErrored, setIsErrored] = useState(false)
+
+	return (
+		<div
+			style={{
+				display: 'grid',
+				gridTemplateColumns: '20rem',
+				gridGap: '1rem',
+				padding: '2rem',
+			}}
+		>
+			<ErrorBoundary>
+				<button
+					onClick={() => {
+						throw new Error('Threw client-side Error')
+					}}
+				>
+					Throw client-side onClick error
+				</button>
+
+				<ThrowerOfErrors
+					isErrored={isErrored}
+					setIsErrored={setIsErrored}
+				/>
+				<button onClick={() => setIsErrored(true)}>
+					Trigger error boundary
+				</button>
+				<button
+					onClick={async () => {
+						throw new Error('an async error occurred')
+					}}
+				>
+					Trigger promise error
+				</button>
+			</ErrorBoundary>
+		</div>
+	)
+}
+
+function ThrowerOfErrors({
+	isErrored,
+	setIsErrored,
+}: {
+	isErrored: boolean
+	setIsErrored: (isErrored: boolean) => void
+}) {
+	useEffect(() => {
+		if (isErrored) {
+			setIsErrored(false)
+			throw new Error('Threw useEffect error')
+		}
+	}, [isErrored, setIsErrored])
+
+	return null
+}
+
+```
+
+## Next Steps
+
+- [Api Instrumentation (Page Router)](./4_api-page-router.md)
+- [Api Instrumentation (App Router)](./5_api-app-router.md)
+- [Edge Runtime](./6_edge-runtime.md)
+- [SSR Error Handlers](./7_ssr-error-handlers.md)
+- [Configuration](./8_configuration.md)
