+++ docs-content/getting-started/fullstack-frameworks/next-js/index.md
@@ -1,0 +1,6 @@
+---
+title: Next.js
+slug: next-js
+createdAt: 2021-09-13T22:07:04.000Z
+updatedAt: 2022-04-01T19:52:59.000Z
+---
