+++ e2e/nextjs/pages/api/page-router-edge-test.ts
@@ -1,7 +1,7 @@
 // pages/api/test.ts
 import { NextRequest } from 'next/server'
 
-import { withEdgeHighlight } from '@/app/utils/edge-highlight.config'
+import { withEdgeHighlight } from '@/app/_utils/edge-highlight.config'
 import { z } from 'zod'
 
 export default withEdgeHighlight(async function handler(request: NextRequest) {
