+++ e2e/nextjs/src/app/components/trpc-queries.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { trpc } from '@/app/utils/trpc'
+import { trpc } from '@/app/_utils/trpc'
 
 export const TrpcQueries = trpc.withTRPC(function TrpcQueries() {
 	const helloWorld = trpc.helloWorld.useQuery({ message: 'Hello Chris' })
